Factored out the relay server into a separate class.
This CL does not factor out more code yet into individual header files
nor does it change the logic of the server or the underlying logging
library. All these changes are necessary but planned for later.
